requirements.txt
fastapi==0.88.0
uvicorn==0.20.0  # uvicorn[standart]
python-multipart==0.0.5
sqlalchemy==1.4.46
email-validator==1.3.1
python-dotenv==0.21.1
flake8==6.0.0

python-jose==3.3.0  # python-jose[cryptography]
cryptography==39.0.0

passlib==1.7.4  # passlib[bcrypt]
bcrypt==4.0.1

# For migrations, and migrations tests
alembic==1.9.2
sqlalchemy_utils==0.39.0
pytest==7.2.1
pytest-env==0.8.1
pytest-order==1.0.1

celery==5.2.7
Redis==4.4.2
httpx==0.23.3
Jinja2==3.1.2
